requirements.txt
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
streamlit==1.20.0
pandas==1.4.2
opencv-python==4.6.0.66
numpy==1.21.5
matplotlib==3.5.1
Pillow==9.0.1
streamlit-image-select==0.5.1